extract client menu method
