+++ generators/base/logic/output/target-path-resolver.mjs
